pyproject.toml
[project]
name = "motyve-us"
version = "0.1.0"
description = "Website for motyveLLC"
requires-python = ">=3.11"
readme = "README.md"
dependencies = [
    "fastapi>=0.135.1",
    "jinja2>=3.1.6",
    "python-dotenv>=1.2.2",
    "uvicorn>=0.41.0",
]

[project.optional-dependencies]
logger = [
    "rich>=14.3.3",
]
